Speed up CPU core by avoiding fetching opcodes every time.
This is a large refactoring in which we modify the Bus interface
to allow fetching a direct memory address to a block. Then we
change the CPU emulation to use instead of the previous opFetch()
calls. To make this valuable, we also need to modify the sync CPU
interface to allow for block runs instead of exposing Step().