geometry: execute commands as soon as they arrive.
Say a command enters a FIFO at time 100 and requires 15 cycles
for its execution. Previously, we were waiting until instant 115
before executing the command and removing it from the FIFO.

Now, as soon as we see it, we execute it right away, and then
increment the cycle counter up to 115. This makes the geometry
FIFO more eager to execute commands and can make end its timeslice
after its target (not a problem anyway).

The effect of this patch is very visible on SwapBuffers, that
can take a long time because it waits for the VBlank.